Add script to export data as CSV
Add a script which exports the test results in CSV format. The results
are normalized so all test results range from 0 to 1.0 where 0 is the
lowest performance result and 1.0 is the highest performance result and
values in between are scaled by the difference between the highest and
lowest performance results. Also, in the cases where the test results
use a smaller number to indicate better performance, the scaled results
are subtracted from 1.0 so the final result will still have the best
performance at 1.0 and the lowest performance at 0. The results are also
filtered to remove some test configurations (currently z840, ZFS, and
8570w) since they would otherwise significantly affect the range,
resulting in a bunch of results being compressed to nearly the same
number.