Add setup-lvm script
Add the setup-lvm script to create the hdd-home logical volume (LV) from
a RAID10 configuration of physical HDDs. Using 4 stripes results in only
using half the capacity for the larger drives (so LVM will use 300 GB of
the 600 GB drives so all 8 drives will be the same size). Using 2
stripes reduces the performance of the RAID10 array but it allows the
full capacity of all the drives to be used.