Add new results
Add results for the case where we use the first SSD as the root drive
and use the second SSD as a LVM cache for a LVM RAID10 drive array
constructed from 8 drives. Both stripes=2 and stripes=4 cases are
provided. In the stripes=4 case, LVM only uses the capacity of the
smallest drive so all drives are 300 GB and the remaining 300 GB on four
of the drives is unused. In the stripes=4 case, LVM uses the full
capacity of each drive but accesses are striped over less drives,
theoretically reducing performance. The stripes=2 case provides the best
overall performance of all of the tested cases.